+++ domain/src/main/java/daily/dayo/domain/usecase/comment/RequestPostCommentUseCase.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class RequestPostCommentUseCase @Inject constructor(
     private val commentRepository: CommentRepository
 ) {
-    suspend operator fun invoke(postId: Int) =
+    suspend operator fun invoke(postId: Long) =
         commentRepository.requestPostComment(postId)
 }
